requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
h5py==3.13.0
iniconfig==2.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
networkx==3.3
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pip==25.0.1
pluggy==1.5.0
pyparsing==3.1.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
scipy==1.15.2
six==1.16.0
pandas==2.2.3
seaborn==0.13.2